+++ pages/accesorios.html
@@ -40,14 +40,16 @@
                 <script src="/scripts/search/searchA.js" type="module"></script>
             </div>
         </div>
+
         <div id="modal">
-            <script src="/scripts/modal.js"></script>
-            <script src="/scripts/carrito.js"></script>
+            <script src="/scripts/carrito/modal.js" type="module"></script>
+            <script src="/scripts/carrito/carrito.js" type="module"></script>
+            <script src="/scripts/carrito/carritoHelpers.js" type="module"></script>
         </div>
     </main>
 
     <footer id="footer">
-        <script src="/scripts/footer.js"></script>
+        <script src="/scripts/footer.js" type="module"></script>
     </footer>
 
     <!-- JS BOOTSTRAP -->
