+++ pages/login.html
@@ -21,7 +21,7 @@
 
 <body class="d-flex flex-column min-vh-100">
     <header id="navContainer">
-        <script src="/scripts/navbar.js"></script>
+        <script src="/scripts/navbar.js" type="module"></script>
     </header>
 
     <main class="flex-fill">
@@ -63,16 +63,17 @@
                 </form>
             </div>
         </div>
-        <script src="/scripts/login.js"></script>
+        <script src="/scripts/login.js" type="module"></script>
 
         <div id="modal">
-            <script src="/scripts/modal.js"></script>
-            <script src="/scripts/carrito.js"></script>
+            <script src="/scripts/carrito/modal.js" type="module"></script>
+            <script src="/scripts/carrito/carrito.js" type="module"></script>
+            <script src="/scripts/carrito/carritoHelpers.js" type="module"></script>
         </div>
     </main>
 
     <footer id="footer">
-        <script src="/scripts/footer.js"></script>
+        <script src="/scripts/footer.js" type="module"></script>
     </footer>
 
     <!-- JS BOOTSTRAP -->
