+++ pages/product.html
@@ -21,24 +21,23 @@
 
 <body class="d-flex flex-column min-vh-100">
     <header id="navContainer">
-        <script src="/scripts/navbar.js"></script>
+        <script src="/scripts/navbar.js" type="module"></script>
     </header>
 
     <main class="flex-fill">
-
         <div id="productDetail">
-            <script type="module" src="/scripts/productDetail.js"></script>
+            <script type="module" src="/scripts/productDetail.js" type="module"></script>
         </div>
 
         <div id="modal">
-            <script src="/scripts/modal.js"></script>
-            <script src="/scripts/carrito.js"></script>
+            <script src="/scripts/carrito/modal.js" type="module"></script>
+            <script src="/scripts/carrito/carrito.js" type="module"></script>
+            <script src="/scripts/carrito/carritoHelpers.js" type="module"></script>
         </div>
-
     </main>
 
     <footer id="footer">
-        <script src="/scripts/footer.js"></script>
+        <script src="/scripts/footer.js" type="module"></script>
     </footer>
 
     <!-- JS BOOTSTRAP -->
