+++ pages/register.html
@@ -21,7 +21,7 @@
 
 <body>
     <header id="navContainer">
-        <script src="/scripts/navBar.js"></script>
+        <script src="/scripts/navBar.js" type="module"></script>
     </header>
 
     <main>
@@ -101,18 +101,19 @@
                     </div>
 
                 </form>
-                <script src="/scripts/register.js"></script>
             </div>
         </div>
+        <script src="/scripts/register.js" type="module"></script>
 
         <div id="modal">
-            <script src="/scripts/modal.js"></script>
-            <script src="/scripts/carrito.js"></script>
+            <script src="/scripts/carrito/modal.js" type="module"></script>
+            <script src="/scripts/carrito/carrito.js" type="module"></script>
+            <script src="/scripts/carrito/carritoHelpers.js" type="module"></script>
         </div>
     </main>
 
     <footer id="footer">
-        <script src="/scripts/footer.js"></script>
+        <script src="/scripts/footer.js" type="module"></script>
     </footer>
 
     <!-- JS BOOTSTRAP -->
