+++ pages/zapatilla.html
@@ -24,7 +24,7 @@
 <body>
     <header>
         <div id="navContainer">
-            <script src="/scripts/navbar.js"></script>
+            <script src="/scripts/navbar.js" type="module"></script>
         </div>
     </header>
 
@@ -42,13 +42,14 @@
         </div>
 
         <div id="modal">
-            <script src="/scripts/modal.js"></script>
-            <script src="/scripts/carrito.js"></script>
+            <script src="/scripts/carrito/modal.js" type="module"></script>
+            <script src="/scripts/carrito/carrito.js" type="module"></script>
+            <script src="/scripts/carrito/carritoHelpers.js" type="module"></script>
         </div>
     </main>
 
     <footer id="footer">
-        <script src="/scripts/footer.js"></script>
+        <script src="/scripts/footer.js" type="module"></script>
     </footer>
 
     <!-- JS BOOTSTRAP -->
