+++ scripts/search/searchA.js
@@ -53,7 +53,7 @@ function renderCards(products) {
                         </div>
                     </div>
                     <p class="text-success fw-bold">$${product.precio.toLocaleString()}</p>
-                    <button class="btn btn-outline-primary">Ver producto</button>
+                    <a href="product.html?id=${product.id}" class="btn btn-outline-primary">Ver producto</a>
                 </div>
             </div>
         </div>`;
