+++ scripts/search/searchI.js
@@ -54,7 +54,7 @@ function renderCards(products) {
                         </div>
                     </div>
                     <p class="text-success fw-bold">$${product.precio.toLocaleString()}</p>
-                    <button class="btn btn-outline-primary">Ver producto</button>
+                    <a href="product.html?id=${product.id}" class="btn btn-outline-primary">Ver producto</a>
                 </div>
             </div>
         </div>`;
